1367749 - Bugfix for OSE default data
* Starting a deployment without OSE, backing up to product selection and
then enabling OSE resulted in default settings never being set for OSE.
This should set default data on the deployment for both OSE and CFME
regardless of if the product is being deployed or not. This shouldn't
have any effect if the products are never deployed because of their
respective guards, deploy_* on the model.