1373432 - 2nd deployment causes unauth log msg
The second deployment has a cached value that it tries to verify is still valid
by hitting an api on the backend server. The original api would throw an
unauthorized error if the portal credentials were not stored on the backend
session. This caused the backend to log an error making it seem like things
failed.

In reality, it was working as expected, the UI checked to see if we were
logged into the portal. Backened responed with an error message, the UI then set
their cached value of isAuthenticated to false.

This behavior while working as expected resulted in a really ugly log message
and an inefficient call.

SOLUTION:
Add an is_authenticated route on the backend so the UI can check to see if the
credentials are stored in the session, thus validating their credentials are
still active. The new api will return true or false. No error message is logged.

Update:

* fix incorrect boolean logic, added tests as well
* fix assert statements